Add page updates
Reformatted add page header
Passed the name to use in the add page for clarity.
Fixed the nav bar for add/edit/view to only display edit/view past the first page (list of site-networks) and add on all but the last (component)
Fixed add page to work on the first page.